Replaces Machete Scabbard in Smartgunner Vendor with new SG scabbard (#6880)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Replaces the normal machete scabbard in the SG vendor with one that can
be placed in the back slot, along with new sprites for it. Price is
increased from 6P to 15P. This scabbard can only be connected to the
harness and will be dropped when the harness is unequipped.

# Explain why it's good for the game
To currently equip a machete as SG, you need to sacrifice your belt. You
can't place it on your back or suit storage. This offers a new but
expensive way to have a machete without sacrificing a belt or pouch
slot. This is specifically a new item so that you can't just equip any
back machete scabbard, you still need to put in some points to get value
out of it.

It also adds some variety to playing SG, as this provides an alternative
to carrying folding cades always as SG.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/118155b7-1ca3-47b8-8858-ea8c92c11db0)

Updated sprite for item

![image](https://github.com/user-attachments/assets/28d47c25-2781-4ec2-b837-980a53a445c4)


</details>


# Changelog
:cl:
add: Smartgunner Machete Scabbard, which smartgunners can wear on their
back when their harness is equipped.
balance: SG's vendor has replaced the 6P Machete Scabbard with a 15P
Smartgunner Machete Scabbard.
/:cl: