IB fixed fun fact (#6863)
# About the pull request

Adds a fun fact for number of times internal bleeding was fixed, similar
to the revives fact.

# Explain why it's good for the game

I'm not a huge fan of fun facts in general as it sometimes incentivizes
bad behavior (such as focusing on revives exclusively and neglecting
other duties just because it has a fun fact), but between removing the
revives fact or just adding a second medical fact, I went this
direction.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: fun fact for internal bleeding fixed
/:cl: