Makes explosions always deal full damage on the source tile (#6888)
# About the pull request
If you jump on the grenade you are going down
(Makes it so you no longer get 50% damage reduction if you are lying
down at the source tile of the explosion)
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fixes some things that were broken (for example sadar rockets or cluster
OB gibbing humans on direct hit) and in general it doesn't make a lot of
sense to get half the damage by hugging the explosion
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Explosions now always deal full damage at the source tile
fix: Fixed sadar rockets not gibbing humans on direct hit
/:cl: