Reduces the Cost of the AGM-184 'Harpoon II' from 300 to 200 (#6876)
# About the pull request

The AGM-184 'Harpoon II' is an underwhelming choice for Gunship Pilots
doing CAS. Running the numbers, it is worse than both the Widowmaker
**and the Banshee** (this one's an incendiary rocket, how is the harpoon
worse than it?), despite costing the same. Below are the comparisons.

Widowmaker | Banshee | **Harpoon**
300 | 175 | **150**
260 | 155 | **134**
220 | 135 | **118**
180 | 115 | **102**
140 | 95 | **86**
100 | 75 | **70**
60 | 55 | **54**
20 | 35 | **38**
0 | 15 | **22**
0 | 0 | **6**

Most notable are the fact that the harpoon deals less damage at every
single tile than the Widowmaker (except for the negligible last three
tiles), and is comparable to the Banshee in strength despite having no
incendiary effect attached to it - and yet, costing the same.

With the change to the cost going from 300 to 200, you can get 6
Harpoons instead of 4 Widowmakers.

# Explain why it's good for the game

Gives the AGM-184 'Harpoon II' the love it deserves, renouncing its
status as a noob trap and troll pick, and provides reason as to why it's
an option to print in the first place.


# Testing Photographs and Procedure

N/A

# Changelog
:cl: MarpleJones
balance: Reduces the cost of the AGM-184 'Harpoon II' from 300 to 200.
/:cl:

---------

Co-authored-by: MarplePhonks <marplejones@gmail.com>