ERT  no longer show up in the orbit escape section (#6910)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
ERT factions no longer show up in the escape section.

This issue was caused by #5058

# Explain why it's good for the game
Makes it clearer who has actually escaped and not ERT returning to their
base.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: ERT no longer show up in the orbit escape section
/:cl: