"Dark" HUD indicators adjust with view size (#6851)
# About the pull request

"Dark" HUD indicators are now anchored to the top of the viewport, and
so are positioned appropriately for various view sizes.

Before:

![image](https://github.com/user-attachments/assets/9363d65c-b9e8-45bb-919f-acc80d36590d)
After:

![image](https://github.com/user-attachments/assets/65eff05a-cf51-42d1-9636-24e1f86b9ba3)

Before, binocs:

![image](https://github.com/user-attachments/assets/cbbbfeb7-4e14-444e-97b7-62f3f56be8ae)
After, binocs:

![image](https://github.com/user-attachments/assets/fe3dc17f-f2ff-40d3-ac79-a94ab7af3732)

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Looks better.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
ui: dark HUD indicators stay anchored to the viewport edge
/:cl: