Keycard authentication device now reqires two users to activate events (#6903)
# About the pull request

it is writen down that two are reqired to activate red and so on but one
can swipe twice fast and it work. this fixes it

# Explain why it's good for the game

bugfix and has rp incentive to get second man to start redalert


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: two persons are needed to activate red, ert and emergency acces
using cards
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>