Corporate Liaison's Wardrobe (#6936)
# About the pull request

This PR adds a cool new vendor for the CL. Instead of that old
overflowing wardrobe full of clothes, this adds a proper vendor that
lets you pick and choose your outfits from a menu rather than
right-clicking a stack. The vendor includes more options than the
cabinet had, including workwear (for business casual), a new orange
outfit I made, civilian versions of the white and black berets, and much
more. All corporate roles can use it, that being any of them listed in
the wy.dm file. Survivors can also use it, as they like to get fresh
sets of clothes.

You're able to get one set of each clothing available, so you can't get
multiples of the same clothes. However, you're able to get ONE of every
item available, so you can easily supply others with clothes. This is
intentional, because certain groups like fresh clothes. This could be
survivors, outside WY visitors, or just someone on the ship who wants to
wear a bomber jacket.

I also did some minor changes to some minor items I had included with
the vendor, such as making a unique variant of the corporate boots
without the knife (making sure all the roles that used to spawn with it
keep the knife version), renaming the 'sunglasses' (hipster glasses) to
proper prescription glasses (considering it's more accurate and we
already have 'sunglasses'), changed the item sprites for some of the
suits (they were old and didn't match their mobsprites), and a bit more.

This PR replaces the CL's cabinet with said vendor, I used the mapmerger
tool, and I tested the mapping commit.

I also tested everything, and it all works ingame.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The current cabinet is overflowing with clothes, and it doesn't leave
any room for new fashion items to be added. The liaison is also one of
the only roles without a vendor when they really should _have_ a
clothing vendor. CC has a vendor, CO has a vendor, CMO has a vendor,
doctors have vendors, nurses have vendors, etc. It's also just a nice
quality of life to be able to select your outfit from a proper GUI,
rather than the right click context menu.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/56cdec25-acec-42f8-91b1-74ed1b081372)
![Screenshot 2024-08-11
140435](https://github.com/user-attachments/assets/d51ffddc-9e73-4552-9756-3340ba530df9)

![image](https://github.com/user-attachments/assets/fccf214b-307e-4844-8d6a-1d093b0fa36b)
![Screenshot 2024-08-11
140453](https://github.com/user-attachments/assets/ad58642f-d14e-4266-876c-27ced2a26e2c)

![image](https://github.com/user-attachments/assets/02e93db9-0bea-4347-bacb-b6d01bbf5c70)

![image](https://github.com/user-attachments/assets/1dc38880-b592-41da-85b8-d36e390115c9)
![Screenshot 2024-08-11
131427](https://github.com/user-attachments/assets/cd356b07-04f2-4f48-9c04-9f9d69199d64)
![Screenshot 2024-08-11
132445](https://github.com/user-attachments/assets/ad8e4e6e-646e-4b49-b6df-1d3d0b9270fc)
![Screenshot 2024-08-11
131712](https://github.com/user-attachments/assets/f81965cc-71e6-4e37-a9c0-7230b9af1a37)
![Screenshot 2024-08-11
143951](https://github.com/user-attachments/assets/7eae84fa-a2d5-4e2c-893c-afc5e93cfcd4)

</details>


# Changelog
:cl: Vandujr
add: Added a new 'Corporate Liaison's Wardrobe' that all corporate roles
can use. includes more clothes than the cabinet originally had.
add: Added a new orange outfit the CL can pick from his Wardrobe Vendor.
add: Added civilian black and white berets.
refactor: Refactored the corporate boots, making a version with and
without a knife in it on spawn. All roles that should have the knife
version, have it.
imageadd: Added new item sprites to the old 'complete' suits; that being
the tan and formal suits, aswell as the suspenders.
imageadd: Added a new item sprite for the black beret, in line with the
rest of the modern berets.
maptweak: Swapped the CL's cabinet for the wardrobe vendor.
/:cl: