Chat Filter (#6901)
# About the pull request
Adds a chat filter which automatically prevents messages in-character,
OOC, LOOC or dchat from being sent if they contain certain words, this
will notify the user and warn them that they will be banned should they
try to circumvent it.

This is obviously just a config option so if this is merged the config
will have to be edited accordingly.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Instead of instantly banning people for saying certain words, simply
prevents them from saying them in the first place - this way no harm is
done to anyone who might be offended as well as the player which
(possibly mistakenly) said the word not being banned and instead let
know the word is not allowed.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
config: Added config options for filtering out certain words
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>