UPP Logistics Worker and Political Commissar Addition (#6920)
# About the pull request

This PR adds two new UPP roles for admin events, the UPP Logistics
Technician and the UPP Political Commissar.

The duties of the Logistics Technician are self-evident, however the
duties of the Political Commissar are ill defined. This is intentional
and it is up to a combination of the staff running the event, and the
player, how they want to have the Commissar interact with the UPP
Command Staff, or USCM forces. They are technically considered a
non-combat military personnel, and are not a civilian.

The PR also adds a few bits of organization to the UPP gear preset file.

# Explain why it's good for the game

Adds more options for admins in events. The Logistics Worker can assist
in managning the UPP supply depot/s whereas the Political Officer can
add unique opportunities for roleplay interactions not typically seen.


# Testing Photographs and Procedure

![Screenshot 2024-08-11 14 56
17](https://github.com/user-attachments/assets/e403a058-0c8d-472f-a2f2-d11750a06419)

Left, Logistics Worker
Right, Commissar

# Changelog
:cl:
add: The UPP Logistics Technician has been added as a event role for the
UPP. Their intention is to assist in operating any supply depots and
logistics departments under the control of the UPP, they also possess
engineering and construction skills allowing them to assist in
fortification work as required.
add: The UPP Political Commissar has been added as a event role for the
UPP. They act to ensure the political ideology of the UPP is upheld
within the unit they are attached to. The goals and ranking of this job
are left undefined to add ambiguity as to their ability to influence the
UPP command structure.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>