Re-adds peridaxon revival mix as a separate option (#6914)
# About the pull request
Re-adds the peridaxon revival mix (peri instead of tricord) with
slightly different ratios (instead of 5:5:5 instead of 6:6:3) as a
loadout option.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Some people prefer their revival mix with peri, some prefer it with
tricord - each having their own ups and downs. You shouldn't be limited
to one or the other (or hoping the chemist feels nice enough to change
the contents of your pressurized canister today).
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added another revival mix pouch option which contains Peridaxon
instead of Tricordrazine.
/:cl: