Re-orders the SEA vendor (#6983)
# About the pull request

As title, changed up the order of items in the SEA vendor a bit, and
changed what some of their internal categories are.

# Explain why it's good for the game

Efficiency


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Re-ordered the SEA vendor contents slightly, and changed the combat
gear to use combat categories.
/:cl: