Fixes teslas in fiorina (#6900)
# About the pull request
Closes #1090 
Makes the tesla on the scavenger ship unhackable for consistency as well
as not shoot the sentry right next to it due to how crowded the space is
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fixes a bug
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed the premade tesla coils on Fiorina

/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>