A few paradrop-related tips. (#6997)
# About the pull request

adds 4 tips related to the art of falling with style
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The only way I figured out a few of these tips was from codediving,
better to keep them in so that people who don't codedive can actually
see it and take advantage of it
# Testing Photographs and Procedure
tested and made sure that the tips are true (Tossed some stuff out while
paradropping and saw what stuck the landing and what got deleted)

# Changelog
:cl:
add: 4 new tips related to paradropping!
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>