Emissive APCs (#7000)
# About the pull request
Adds emissive lights to APCs, Hopefully this PR shows people how easy it
is to add emissives so we can make the game prettier

![image](https://github.com/user-attachments/assets/e0f3c9ad-1a1e-47be-9fd9-e26851dfa34a)

![image](https://github.com/user-attachments/assets/8707353b-beab-4a1b-98d2-4fa956bd84c3)

![image](https://github.com/user-attachments/assets/3a7f65f2-aed0-42aa-8b7a-ffa9bd14f6b1)
Takes inspiration from
https://github.com/tgstation/TerraGov-Marine-Corps/pull/14257 and
https://github.com/tgstation/TerraGov-Marine-Corps/pull/6739 and
https://github.com/cmss13-devs/cmss13/pull/5405
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it prettier
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: APCs are now emissive
code: Added emissive blockers to a bunch of stuff
/:cl: