Smartgun battery name sanity (#6934)
# About the pull request
Moves a word to the right abit so it looks like every other item with
"letter_number_designation item_name_here"
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Saves my sanity a little?
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
spellcheck: places smartgun after DV9 to be DV9 smartgun battery
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>