small vai tweak (#6819)
# About the pull request

Adds a tactical HPR to the VAIPO and VAISO ERTs, VAISO MG has it 100% of
the time, VAIPO 33%, the tact-HPR has an agrip, magharn, suppressor, and
has a 2 damage mult to bring its dmg up to M41A/MAR50 levels
Adds an HPR AP mag, just in vai ert gear, 3 mags for vaipo(incl loaded)
and 4 for vaiso(incl loaded)
Buffs the MAR50 fire rate a bit to lmg levels.

oh also since I was in the mag code I noticed the formatting on the hpr
heap mag was messed up so I changed it.

# Explain why it's good for the game

tact-HPR fits vaiso more and I've been meaning to do this since I added
the ERT; I also tossed it to vaipo too, the dmg boost is to make it
on-par with the m41 as they do not get a bipod to take advantage of
bipoded ROF, as well as making them capable in mobile situations.
AP mags are to fit in with the rest of the gear everyone else has, might
be a bridge too far but I can simply remove it if needed.
mar50 shoots really slowly when not bipoded or on burst, to the point of
uselessness.


Honestly I feel HPR should simply get that damage boost full stop and
simply have it removed when bipoding but that would be annoying to code.



# Changelog
:cl:
add: Adds a tactical HPR to VAISO and VAIPO machinegunners
add: Adds an HPR AP mag available to the above
balance: MAR50 has normal LMG level firerate now
spellcheck: HPR HEAP mag now has correct name formatting
/:cl: