New sound for revive notification (#6918)
# About the pull request

The sound to notify you that you're being revived is no longer the ahelp
sound, now it's a modification of the defib noise.

# Explain why it's good for the game

More sound diversity (and it's an existing sound anyways just my own
edit to it)


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
soundadd: revive notification has new sound (rather than the ahelp
sound)
/:cl: