Solar Devils ERT [PvE Marines] (#6811)
# About the pull request

In marking the merger of the PvP and PvE servers under a single banner,
this PR sets to celebrate that by the addition of the 'Solar Devils'
battalion as a (currently admin only) Emergency Response Team. The Solar
Devils are similarly well equipped and trained as they appear in PvE.

This PR also adds some content from the PvE server to enhance the
appearance of the Solar Devils to better fit their appearance, such as a
battalion patch sprite.

These squads will spawn with, depending on if they are a half or full
squad, 4 Riflemen, 2 Smartgunners, 2 Team Leaders, 1 Platoon Corpsman
and 1 Platoon Leader. The Platoon Leader will only spawn with the 'Full
Squad' ERT.

The squad are all armed with M41A Mk1's.

----

In terms of function. This PR adds two distress calls for admins to use,
a Half-Squad and Full Squad deployment of either 5 or 10, respectively,
Solar Devil Marines. They spawn with some unique equipment, and they all
possess increased skills.

# Explain why it's good for the game

Show of comradeship, and a small easter egg of the PvE server in PvP. 


# Testing Photographs and Procedure
![Screenshot 2024-07-28 21 52
07](https://github.com/user-attachments/assets/a28fca30-a2ca-4b13-9528-42586d4a8caf)



# Changelog
:cl:
add: In observance of the merger of the PvP and PvE servers of CM-SS13.
The 'Solar Devils', the USCM battalion that appear in the PvE server,
have been added as an admin only Emergency Response Team.
add: The Solar Devils are highly trained and well equipped by Marine
standards, they will also spawn as part of their own squad.
imageadd: A 'Solar Devils' uniform patch accessory has been added. 
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: private-tristan <54422837+private-tristan@users.noreply.github.com>