Bursting+facehugger stats no longer count NPCs (#6978)
# About the pull request

Bursting+facehugger stats no longer count NPCs

# Explain why it's good for the game

Right now the stats seem to be half and half on whether they count NPCs
or not, it should be one way or the other.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: bursting+facehugger stats disregard NPCs
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>