Req Vendors Typo/Clarity/Remove RO mentions (#7036)
# About the pull request
Addresses #7033
Adds an s to Corpman to make Corpsman
Murders RO in place for QM
Extends the description of the gun vendor to also include that it has
all the storage thingies and the misc stuff.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
More clarity, less old and dead RO.
Plus a typo fix.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas

spellcheck: Req vendor descs remove RO. Fix Corpsman Typo
/:cl: