Xenos no more killing crit allies by dragging/devouring them (#7044)
# About the pull request

Currently, xenos allied to a faction or an individual person will
accidentally kill them if they happen to drag or devour them when they
are in critical state. This PR seeks to fix that.

# Explain why it's good for the game

It's great not accidentally killing your allies when you are trying to
save them!

# Testing Photographs and Procedure

Just tested it with allied factions and individuals, both work! 

Enemies of the hive still die like usual if they are dragged/devoured in
crit (as the poor CLF showcases)

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/bff213a2-d8b7-4832-9ab5-578c1b3b40da)

</details>


# Changelog

:cl: SubjectD9341

add: xenos don't kill crit allies by dragging/devouring anymore

/:cl: